Enable fractional variable difficulty (vardiff) support
This commit enables support for variable difficulty values below 1.0,
allowing the pool to manage low-hashrate miners (ESP32, etc.) with
fine-grained difficulty adjustment.

Changes:
1. src/stratifier.c:
   - Changed worker->mindiff from int to double (line 197)
   - Changed local mindiff variables from int64_t to double (lines 5677, 7779)
   - Replaced lround() with direct double assignment in optimal calculation (lines 5758, 5760)
   - Changed vardiff floor check from 'optimal < 1' to 'optimal <= 0.0' (line 5773)

2. src/ckpool.c:
   - Added config validation for mindiff (hard floor at 0.0, soft warning at 0.001)
   - Changed json_get_int64 to json_get_double for highdiff/maxdiff parsing

3. src/ckpool.h:
   - Changed highdiff and maxdiff from int64_t to double for consistency

4. tests/:
   - Added test-fractional-stats.c: Unit tests for fractional share accumulation
   - Added test-fractional-vardiff.c: Unit tests for fractional vardiff algorithm
   - Updated Makefile.am to include new test programs

Testing & Validation:
- All 21 unit tests passing (including 2 new fractional tests)
- Live testing with multiple miners (low-hashrate CPU, high-hashrate GH/s)
- Vardiff adjusted smoothly from 0.001 to 464+ across full range
- No oscillation, clean exponential convergence to equilibrium
- 1.17M+ shares processed with excellent acceptance rate
- Hashrate statistics accurate and stable across all time windows

This feature enables proper support for fractional difficulties in both
directions (up and down), allowing optimal difficulty targeting for miners
across the entire hashrate spectrum.

+++ src/ckpool.h
@@ -239,8 +239,8 @@ struct ckpool_instance {
 	/* Difficulty settings */
 	double mindiff; // Default 1.0
 	double startdiff; // Default 42.0
-	int64_t highdiff; // Default 1000000
-	int64_t maxdiff; // No default
+	double highdiff; // Default 1000000
+	double maxdiff; // No default
 	bool allow_low_diff; // Allow network diff below 1.0 (for regtest testing)
 
 	/* Coinbase data */
